Ignore remote dbname when client requested default
When the client not not request an explicit database name, the server
may set the database parameter on the relation to the actual database
name to connect to, and not mirror back the request identically (ie.
set it to the empty string). This allows legacy clients using the
legacy relation fields to still construct their own connection strings
to the database.